// src/types.ts
export interface Pillar {
  title: string;
  quote: string;
  description: string;
}

export interface TeamMember {
  name: string;
  role: string;
  bio: string;
  image: string;
}

export interface Blog {
  title: string;
  author: string;
  date: string;
  readTime: string;
  image: string;
  content: string;
}